requirements.txt
streamlit
requests
reportlab
python-docx
